gralloc: uncompressed format doesn't need gpu realloc [1/3]
PD#SWPL-193049

Problem:
after modifying the nv12 format, the h264 process goes to
the ge2d copy, but the ge2d is not adapted to nv12, and
an error occurs during data copying.

Solution:
a compress member is added to decoder_para to determine
whether GPU reallocation is required.

Verify:
android-u

Change-Id: I16831d7add16a49d0111168417e1f74ab1c46264
Signed-off-by: yang.zhou <yang.zhou@amlogic.com>